♻️refactor: move components to _components folders
Restructure the project by moving component files to dedicated _components
folders within their respective feature directories. This change follows Next.js
conventions for better organization and maintainability.
