+++ src/app/(auth)/sign-in/_components/signin_form.tsx
@@ -4,7 +4,7 @@ import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { ActionState } from '@/types';
 import { useActionState, useRef } from 'react';
-import { signInAction } from '../_actions';
+import { signInAction } from '../../_actions';
 import { Label } from '@/components/ui/label';
 
 export default function SignInForm() {
