+++ src/app/(core)/exercises/_components/exercise-list.tsx
