+++ src/app/(core)/plan/_components/plan_view.tsx
